Effectful.Concurrent.Async

Description

Lifted Control.Concurrent.Async.

Effect

data Concurrent :: Effect

Provide the ability to run Eff computations concurrently in multiple threads and communicate between them.

Warning: unless you stick to high level functions from the withAsync family, the Concurrent effect makes it possible to escape the scope of any scoped effect operation. Consider the following:

>>> import qualified Effectful.Reader.Static as R

>>> printAsk msg = liftIO . putStrLn . (msg ++) . (": " ++) =<< R.ask

>>> :{
  runEff . R.runReader "GLOBAL" . runConcurrent $ do
    a <- R.local (const "LOCAL") $ do
      a <- async $ do
        printAsk "child (first)"
        threadDelay 20000
        printAsk "child (second)"
      threadDelay 10000
      printAsk "parent (inside)"
      pure a
    printAsk "parent (outside)"
    wait a
:}
child (first): LOCAL
parent (inside): LOCAL
parent (outside): GLOBAL
child (second): LOCAL

Note that the asynchronous computation doesn't respect the scope of local, i.e. the child thread still behaves like it's inside the local block, even though the parent thread already got out of it.

This is because the value provided by the Reader effect is thread local, i.e. each thread manages its own version of it. For the Reader it is the only reasonable behavior, it wouldn't be very useful if its "read only" value was affected by calls to local from its parent or child threads.

However, the cut isn't so clear if it comes to effects that provide access to a mutable state. That's why statically dispatched State and Writer effects come in two flavors, local and shared:

>>> import qualified Effectful.State.Static.Local as SL
>>> :{
  runEff . SL.execState "Hi" . runConcurrent $ do
    replicateConcurrently_ 3 $ SL.modify (++ "!")
:}
"Hi"

>>> import qualified Effectful.State.Static.Shared as SS
>>> :{
  runEff . SS.execState "Hi" . runConcurrent $ do
    replicateConcurrently_ 3 $ SS.modify (++ "!")
:}
"Hi!!!"

In the first example state updates made concurrently are not reflected in the parent thread because the value is thread local, but in the second example they are, because the value is shared.

Handlers

runConcurrent :: IOE :> es => Eff (Concurrent : es) a -> Eff es a

Run the Concurrent effect.

Asynchronous actions

data Async a

An asynchronous action spawned by async or withAsync. Asynchronous actions are executed in a separate thread, and operations are provided for waiting for asynchronous actions to complete and obtaining their results (see e.g. wait).

High-level API

Spawning with automatic cancelation

withAsync :: Concurrent :> es => Eff es a -> (Async a -> Eff es b) -> Eff es b

Lifted withAsync.

withAsyncBound :: Concurrent :> es => Eff es a -> (Async a -> Eff es b) -> Eff es b

Lifted withAsyncBound.

withAsyncOn :: Concurrent :> es => Int -> Eff es a -> (Async a -> Eff es b) -> Eff es b

Lifted withAsyncOn.

withAsyncWithUnmask :: Concurrent :> es => ((forall c. Eff es c -> Eff es c) -> Eff es a) -> (Async a -> Eff es b) -> Eff es b

Lifted withAsyncWithUnmask.

withAsyncOnWithUnmask :: Concurrent :> es => Int -> ((forall c. Eff es c -> Eff es c) -> Eff es a) -> (Async a -> Eff es b) -> Eff es b

Lifted withAsyncOnWithUnmask.

Querying Asyncs

wait :: Concurrent :> es => Async a -> Eff es a

Lifted wait.

poll :: Concurrent :> es => Async a -> Eff es (Maybe (Either SomeException a))

Lifted poll.

waitCatch :: Concurrent :> es => Async a -> Eff es (Either SomeException a)

Lifted waitCatch.

asyncThreadId :: Async a -> ThreadId

Returns the ThreadId of the thread running the given Async.

cancel :: Concurrent :> es => Async a -> Eff es ()

Lifted cancel.

uninterruptibleCancel :: Concurrent :> es => Async a -> Eff es ()

Lifted uninterruptibleCancel.

cancelWith :: (Exception e, Concurrent :> es) => Async a -> e -> Eff es ()

Lifted cancelWith.

data AsyncCancelled

The exception thrown by cancel to terminate a thread.

Constructors

AsyncCancelled

compareAsyncs :: Async a -> Async b -> Ordering

Compare two Asyncs that may have different types by their ThreadId.

High-level utilities

race :: Concurrent :> es => Eff es a -> Eff es b -> Eff es (Either a b)

Lifted race.

race_ :: Concurrent :> es => Eff es a -> Eff es b -> Eff es ()

Lifted race_.

concurrently :: Concurrent :> es => Eff es a -> Eff es b -> Eff es (a, b)

Lifted concurrently.

concurrently_ :: Concurrent :> es => Eff es a -> Eff es b -> Eff es ()

Lifted concurrently_.

mapConcurrently :: (Traversable f, Concurrent :> es) => (a -> Eff es b) -> f a -> Eff es (f b)

Lifted mapConcurrently.

forConcurrently :: (Traversable f, Concurrent :> es) => f a -> (a -> Eff es b) -> Eff es (f b)

Lifted forConcurrently.

mapConcurrently_ :: (Foldable f, Concurrent :> es) => (a -> Eff es b) -> f a -> Eff es ()

Lifted mapConcurrently_.

forConcurrently_ :: (Foldable f, Concurrent :> es) => f a -> (a -> Eff es b) -> Eff es ()

Lifted forConcurrently_.

replicateConcurrently :: Concurrent :> es => Int -> Eff es a -> Eff es [a]

Lifted replicateConcurrently.

replicateConcurrently_ :: Concurrent :> es => Int -> Eff es a -> Eff es ()

Lifted replicateConcurrently_.

Concurrently

newtype Concurrently es a

Lifted Concurrently.

Constructors

Concurrently

Fields

runConcurrently :: Eff es a

Conc

data Conc :: [Effect] -> Type -> Type

Lifted Conc.

conc :: Eff es a -> Conc es a

Lifted conc.

runConc :: Concurrent :> es => Conc es a -> Eff es a

Lifted runConc.

data ConcException

Things that can go wrong in the structure of a Conc. These are programmer errors.

Since: unliftio-0.2.9.0

Constructors

EmptyWithNoAlternative

Pooled concurrency

pooledMapConcurrentlyN :: (Concurrent :> es, Traversable t) => Int -> (a -> Eff es b) -> t a -> Eff es (t b)

Lifted pooledMapConcurrentlyN.

pooledMapConcurrently :: (Concurrent :> es, Traversable t) => (a -> Eff es b) -> t a -> Eff es (t b)

Lifted pooledMapConcurrently.

pooledMapConcurrentlyN_ :: (Concurrent :> es, Foldable f) => Int -> (a -> Eff es b) -> f a -> Eff es ()

Lifted pooledMapConcurrentlyN.

pooledMapConcurrently_ :: (Concurrent :> es, Foldable f) => (a -> Eff es b) -> f a -> Eff es ()

Lifted pooledMapConcurrently_.

pooledForConcurrentlyN :: (Concurrent :> es, Traversable t) => Int -> t a -> (a -> Eff es b) -> Eff es (t b)

Lifted pooledForConcurrentlyN.

pooledForConcurrently :: (Concurrent :> es, Traversable t) => t a -> (a -> Eff es b) -> Eff es (t b)

Lifted pooledForConcurrently.

pooledForConcurrentlyN_ :: (Concurrent :> es, Foldable f) => Int -> f a -> (a -> Eff es b) -> Eff es ()

Lifted pooledForConcurrentlyN.

pooledForConcurrently_ :: (Concurrent :> es, Foldable f) => f a -> (a -> Eff es b) -> Eff es ()

Lifted pooledForConcurrently_.

pooledReplicateConcurrentlyN :: Concurrent :> es => Int -> Int -> Eff es a -> Eff es [a]

Lifted pooledReplicateConcurrentlyN.

pooledReplicateConcurrently :: Concurrent :> es => Int -> Eff es a -> Eff es [a]

Lifted pooledReplicateConcurrently.

pooledReplicateConcurrentlyN_ :: Concurrent :> es => Int -> Int -> Eff es a -> Eff es ()

Lifted pooledReplicateConcurrentlyN_.

pooledReplicateConcurrently_ :: Concurrent :> es => Int -> Eff es a -> Eff es ()

Lifted pooledReplicateConcurrently_.

Specialised operations

STM operations

waitSTM :: Async a -> STM a

A version of wait that can be used inside an STM transaction.

pollSTM :: Async a -> STM (Maybe (Either SomeException a))

A version of poll that can be used inside an STM transaction.

waitCatchSTM :: Async a -> STM (Either SomeException a)

A version of waitCatch that can be used inside an STM transaction.

Waiting for multiple Asyncs

waitAny :: Concurrent :> es => [Async a] -> Eff es (Async a, a)

Lifted waitAny.

waitAnyCatch :: Concurrent :> es => [Async a] -> Eff es (Async a, Either SomeException a)

Lifted waitAnyCatch.

waitAnyCancel :: Concurrent :> es => [Async a] -> Eff es (Async a, a)

Lifted waitAnyCancel.

waitAnyCatchCancel :: Concurrent :> es => [Async a] -> Eff es (Async a, Either SomeException a)

Lifted waitAnyCatchCancel.

waitEither :: Concurrent :> es => Async a -> Async b -> Eff es (Either a b)

Lifted waitEither.

waitEitherCatch :: Concurrent :> es => Async a -> Async b -> Eff es (Either (Either SomeException a) (Either SomeException b))

Lifted waitEitherCatch.

waitEitherCancel :: Concurrent :> es => Async a -> Async b -> Eff es (Either a b)

Lifted waitEitherCancel.

waitEitherCatchCancel :: Concurrent :> es => Async a -> Async b -> Eff es (Either (Either SomeException a) (Either SomeException b))

Lifted waitEitherCatchCancel.

waitEither_ :: Concurrent :> es => Async a -> Async b -> Eff es ()

Lifted waitEither_.

waitBoth :: Concurrent :> es => Async a -> Async b -> Eff es (a, b)

Lifted waitBoth.

Waiting for multiple Asyncs in STM

waitAnySTM :: [Async a] -> STM (Async a, a)

A version of waitAny that can be used inside an STM transaction.

Since: async-2.1.0

waitAnyCatchSTM :: [Async a] -> STM (Async a, Either SomeException a)

A version of waitAnyCatch that can be used inside an STM transaction.

Since: async-2.1.0

waitEitherSTM :: Async a -> Async b -> STM (Either a b)

A version of waitEither that can be used inside an STM transaction.

Since: async-2.1.0

waitEitherCatchSTM :: Async a -> Async b -> STM (Either (Either SomeException a) (Either SomeException b))

A version of waitEitherCatch that can be used inside an STM transaction.

Since: async-2.1.0

waitEitherSTM_ :: Async a -> Async b -> STM ()

A version of waitEither_ that can be used inside an STM transaction.

Since: async-2.1.0

waitBothSTM :: Async a -> Async b -> STM (a, b)

A version of waitBoth that can be used inside an STM transaction.

Since: async-2.1.0

Low-level API

Spawning (low-level API)

async :: Concurrent :> es => Eff es a -> Eff es (Async a)

Lifted async.

asyncBound :: Concurrent :> es => Eff es a -> Eff es (Async a)

Lifted asyncBound.

asyncOn :: Concurrent :> es => Int -> Eff es a -> Eff es (Async a)

Lifted asyncOn.

asyncWithUnmask :: Concurrent :> es => ((forall b. Eff es b -> Eff es b) -> Eff es a) -> Eff es (Async a)

Lifted asyncWithUnmask.

asyncOnWithUnmask :: Concurrent :> es => Int -> ((forall b. Eff es b -> Eff es b) -> Eff es a) -> Eff es (Async a)

Lifted asyncOnWithUnmask.

Linking

link :: Concurrent :> es => Async a -> Eff es ()

Lifted link.

linkOnly :: Concurrent :> es => (SomeException -> Bool) -> Async a -> Eff es ()

Lifted linkOnly.

link2 :: Concurrent :> es => Async a -> Async b -> Eff es ()

Lifted link2.

link2Only :: Concurrent :> es => (SomeException -> Bool) -> Async a -> Async b -> Eff es ()

Lifted link2Only.

data ExceptionInLinkedThread

Constructors

ExceptionInLinkedThread (Async a) SomeException