Effectful.Console.ByteString

Description

Lifted functions from Data.ByteString.Char8 that are related to standard streams. Like the original module, you probably want to import this module qualified to avoid name clashes with the functions provided by Prelude, e.g.:

import Data.ByteString (ByteString)
import qualified Data.ByteString.Char8 as BS8
import Effectful.Console.ByteString (Console)
import qualified Effectful.Console.ByteString as Console

Effect

data Console :: Effect

An effect for reading from/writing to stdin, stdout or stderr.

Handlers

runConsole :: IOE :> es => Eff (Console : es) a -> Eff es a

Run the Console effect.

Operations

getLine :: Console :> es => Eff es ByteString

Lifted getLine.

getContents :: Console :> es => Eff es ByteString

Lifted getContents.

putStr :: Console :> es => ByteString -> Eff es ()

Lifted putStr.

putStrLn :: Console :> es => ByteString -> Eff es ()

Lifted putStrLn.

interact :: Console :> es => (ByteString -> ByteString) -> Eff es ()

Lifted interact.