Effectful.FileSystem.IO.ByteString.Builder

Description

Lifted Data.ByteString.Builder. Like the original module, you probably want to import this module qualified to avoid name clashes with the functions provided by Prelude, e.g.:

import Data.ByteString.Builder (Builder)
import qualified Data.ByteString.Builder as BSB
import qualified Effectful.FileSystem.IO.ByteString.Builder as EBSB

Executing Builders

hPutBuilder :: FileSystem :> es => Handle -> Builder -> Eff es ()

Lifted hPutBuilder.

writeFile :: FileSystem :> es => FilePath -> Builder -> Eff es ()

Lifted writeFile.