Effectful.FileSystem.IO.ByteString.Lazy

Description

Lifted Data.ByteString.Lazy.Char8. Like the original module, you probably want to import this module qualified to avoid name clashes with the functions provided by Prelude, e.g.:

import Data.ByteString.Lazy (ByteString)
import qualified Data.ByteString.Lazy.Char8 as LBS8
import qualified Effectful.FileSystem.IO.ByteString.Lazy as ELBS

Files

readFile :: FileSystem :> es => FilePath -> Eff es ByteString

Lifted readFile.

writeFile :: FileSystem :> es => FilePath -> ByteString -> Eff es ()

Lifted writeFile.

appendFile :: FileSystem :> es => FilePath -> ByteString -> Eff es ()

Lifted appendFile.

I/O with Handles

hGetContents :: FileSystem :> es => Handle -> Eff es ByteString

Lifted hGetContents.

hGet :: FileSystem :> es => Handle -> Int -> Eff es ByteString

Lifted hGet.

hGetNonBlocking :: FileSystem :> es => Handle -> Int -> Eff es ByteString

Lifted hGetNonBlocking.

hPut :: FileSystem :> es => Handle -> ByteString -> Eff es ()

Lifted hPut.

hPutNonBlocking :: FileSystem :> es => Handle -> ByteString -> Eff es ByteString

Lifted hPutNonBlocking.

hPutStr :: FileSystem :> es => Handle -> ByteString -> Eff es ()

Lifted hPutStr.

hPutStrLn :: FileSystem :> es => Handle -> ByteString -> Eff es ()

Lifted hPutStrLn.